requirements.txt
# Dependências para criar app desktop com Python
pywebview==5.1
pyinstaller==6.3.0
